requirements.txt
numpy == 1.16.0
pandas == 0.23.0
tqdm == 4.29.1
jieba == 0.39
re == 2.2.1
matplotlib == 2.1.0
keras == 2.2.4
keras-layer-normalization == 0.11.0
sklearn == 0.20.2
tensorflow-gpu == 1.12.0